handroll storage slot
foo

+++ src/types/EmissaryStatus.sol
@@ -11,7 +11,10 @@ enum EmissaryStatus {
 }
 
 struct EmissaryConfig {
-    address emissary;
-    ResetPeriod resetPeriod;
-    uint48 assignableAt;
+    // 20 bytes
+    address emissary; // address of the sponsor's emissary
+    // 1 byte
+    ResetPeriod resetPeriod; // period after which a set emissary can be re-assigned
+    // 11 bytes
+    uint88 assignableAt; // timestamp after which an emissary can be re-assigned
 }
